+++ example/a.ts
@@ -1,19 +1,16 @@
+// server.ts
 import { Elysia, InternalServerError, t } from '../src'
-import { describe, it, expect, beforeEach } from 'bun:test'
-
-let isOnResponseCalled: boolean
-let onResponseCalledCounter = 0
 
 class CustomError extends Error {}
 
-const app = new Elysia({ aot: true })
-	.onError(() => {
-
+const app = new Elysia()
+	.onAfterResponse((context) => {
+		console.log(context.query)
 	})
-	.onAfterResponse(() => {
-		isOnResponseCalled = true
-		onResponseCalledCounter++
-		console.log(onResponseCalledCounter)
+	.post('/', () => 'yay', {
+		body: t.Object({
+			test: t.String()
+		})
 	})
 	.get('/customError', () => {
 		throw new CustomError('whelp')
@@ -23,53 +20,49 @@ const app = new Elysia({ aot: true })
 	})
 	.listen(3000)
 
-// console.log(app.routes[0].compile().toString())
-// console.log(app.handleError.toString())
+console.log(app.server!.url.toString())
 
-app.handle(new Request('http://localhost/customError'))
-
-// beforeEach(() => {
-// 	isOnResponseCalled = false
-// 	onResponseCalledCounter = 0
-// })
-
+// client.ts
 export const newReq = (params?: {
 	path?: string
 	headers?: Record<string, string>
 	method?: string
 	body?: string
-}) => new Request(`http://localhost${params?.path ?? '/'}`, params)
+	query: string
+}) =>
+	new Request(
+		`http://localhost:3000${params?.path ?? '/'}?name=${params?.query}`,
+		params
+	)
 
-// describe('Error', () => {
-// 	it.each([
-// 		['NotFoundError', newReq({ path: '/notFound' })],
-// 		[
-// 			'ParseError',
-// 			newReq({
-// 				method: 'POST',
-// 				headers: { 'Content-Type': 'application/json' },
-// 				body: ''
-// 			})
-// 		],
-// 		[
-// 			'ValidationError',
-// 			newReq({
-// 				method: 'POST',
-// 				headers: { 'Content-Type': 'application/json' },
-// 				body: JSON.stringify({})
-// 			})
-// 		],
-// 		['CustomError', newReq({ path: '/customError' })],
-// 		['InternalServerError', newReq({ path: '/internalError' })]
-// 	])('%s should call onResponse', async (_name, request) => {
-// 		expect(isOnResponseCalled).toBeFalse()
-// 		expect(onResponseCalledCounter).toBe(0)
-// 		await app.handle(request)
+const cases = [
+	['NotFoundError', newReq({ path: '/notFound', query: 'NotFoundError' })],
+	[
+		'ParseError',
+		newReq({
+			method: 'POST',
+			headers: { 'Content-Type': 'application/json' },
+			body: '',
+			query: 'ParseError'
+		})
+	],
+	[
+		'ValidationError',
+		newReq({
+			method: 'POST',
+			headers: { 'Content-Type': 'application/json' },
+			body: JSON.stringify({}),
+			query: 'ValidationError'
+		})
+	],
+	['CustomError', newReq({ path: '/customError', query: 'CustomError' })],
+	[
+		'InternalServerError',
+		newReq({ path: '/internalError', query: 'InternalServerError' })
+	]
+] as const
 
-// 		// wait for next tick
-// 		await Bun.sleep(1)
-
-// 		expect(isOnResponseCalled).toBeTrue()
-// 		expect(onResponseCalledCounter).toBe(1)
-// 	})
-// })
+for (const [name, req] of cases) {
+	console.log("N", name)
+	await fetch(req)
+}